requirements.txt
asgiref==3.8.1
cloudinary==1.36.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.16
django-allauth==0.57.2
django-crispy-forms==2.3
django-filter==24.3
geographiclib==2.0
geopy==2.4.1
gunicorn==20.1.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.1
urllib3==1.26.20
whitenoise==5.3.0
